fix(crash on crypto search): App crashed when crypto name contained numbers
The app crashed when clicking the add chart button on the search result of a crypto that has at least one number in the name. The reason was that the regex match in the getCryptoData function didn't look for numbers.
Solved by simplyfing the function to not use regex match. Instead, currencyPair.split("/") is called and assigned to the splitCurrency constant. Then, the fromCurrency constant is set to splitCurrency[0], and toCurrency to splitCurrency[1].