test(add unit-, integration-, and e2e-tests):
Added unit-, integration-, and end-to-end-tests. These tests test that DOM elements are displayed correctly, as well as the main functionality of the app, which is to open and close the modal, search for a chart, add the chart to the dashboard, and remove the chart from the dashboard.

In addition to this, I also made these additional, minor changes:

  - In App.tsx: Made the modal be displayed conditionally using the && operator. This makes the tests run correctly, since the moda otherwise is present in the DOM and found using the getByText query. But this also removes the entering and exiting animations from the component. My plan is to solve this in issue in the future by adding the 3rd party library “React Transition Group”.

  - Changed the Modal component to directly return ReactDOM.createPortal, instead of putting it inside the ModalOverlay component first.

  - In the index.html file: Changed the meta title “finDash - A dashboard for stock, currency, and crypto prices” to “finDash - A dashboard for stock, currency, and crypto charts”.