refactor(modal animation):
Changed the way the modal is displayed from being conditionally displayed when isModalShowing is true, using the && operator, to adding CSS classes when isModalShowing is true. This lets the rendering of the modal show an animation.

The reason for switching to the conditional rendering in the first place was so that a test would pass. The jestDom library in React Testing Library wasn't working correctly when expecting an element to be visible, which made me have to find another solution.

I'm now switching back, because I found a way to make the tests pass and show the modal animation. The solution was to add a test-id to the modal, which has a name depending on whether the modal is showing or not.